+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Create `ShoppingCart` class utilizing `IndexedDB`
+
 ## [v2.6.2] - 2021-02-15
 
 ### Added